move Distribution to its own file
